+++ src/main/java/mybook_insight/io/mybook_insight/domain/common/BaseEntity.java
@@ -3,12 +3,15 @@ package mybook_insight.io.mybook_insight.domain.common;
 import jakarta.persistence.Column;
 import jakarta.persistence.MappedSuperclass;
 import jakarta.persistence.PrePersist;
+import lombok.Getter;
+
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
 
 import java.time.LocalDateTime;
 
+@Getter
 @MappedSuperclass
 @SQLDelete(sql = "UPDATE {h-schema}{h-table} SET deleted_at = NOW() WHERE id = ?")
 @SQLRestriction( "deleted_at IS NULL")
