+++ src/test/java/mybook_insight/io/mybook_insight/resources/application-test.yml
@@ -1,0 +1,21 @@
+spring:
+  # datasource
+  datasource:
+    url:
+    username:
+    password:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  # jpa
+  jpa:
+    hibernate:
+      ddl-auto: create
+    show-sql: true
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.MySQL8Dialect
+        format_sql: true
+
+  # redis
+  redis:
+    host: localhost
+    port: 6379
